full on streach to code code now running with simulation manager and all the new classes seem to work rsu wasn't checked and triangulation doenst work yet
